Refactor contracts controller into a Contracts client class.
This provides finer-grained control over waiting for vent via sync helpers and
`do` function. Rather than choosing between synchronising every call you
can synchronise over a lexical/function scope with `do` or you can call
`sync` to perform a sync at any point for a particular contracts
instance (imagined to be request scoped). This is achieved by
maintaining a high-water mark form Burrow returns as well as
Postgres Vent notifications which can be compared.

Along the way cleaned up the contracts code in numerous ways:

- Better encapsulation of initialisation logic
- Use promise API of Burrow everywhere
- Better use of async and promises
- Improved vent helper (handles name reg results)
- Up-front failure on missing environment variables
- Await connections
