+++ api/test/controllers/contracts-test.js
@@ -11,7 +11,7 @@ const crypto = require('crypto');
 const logger = require('../../common/logger');
 const log = logger.getLogger('tests.Harness');
 
-const contracts = require('../../controllers/contracts-controller');
+const contracts = require('../../contracts');
 
 before(function (done) { // eslint-disable-line func-names
   this.timeout(99999999);
